apps/api/pyproject.toml
[project]
name = "neurohub-api"
version = "0.1.0"
description = "NeuroHub API - medical AI workflow orchestrator"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "asyncpg>=0.30.0",
  "alembic>=1.14.0",
  "pydantic-settings>=2.6.0",
  "redis>=5.2.0",
  "celery[redis]>=5.4.0",
  "python-jose[cryptography]>=3.3.0",
  "httpx>=0.28.0",
  "psycopg2-binary>=2.9.0",
  "Pillow>=10.0.0",
  "weasyprint>=62.0",
  "python-multipart>=0.0.9",
  "passlib[bcrypt]>=1.7.4",
  "boto3>=1.35.0",
  "email-validator>=2.0.0",
  "pydicom>=2.4.0",
  "pynetdicom>=2.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.8.0",
  "bandit>=1.7.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
